+++ src/api/api_client.rs
@@ -308,278 +308,8 @@ where
 		Ok(())
 	}
 
-	pub fn get_account_info<AccountId: Clone + Encode>(
-		&self,
-		address: &AccountId,
-	) -> ApiResult<Option<AccountInfo<Runtime::Index, Runtime::AccountData>>> {
-		let storagekey: sp_core::storage::StorageKey =
-			self.metadata
-				.storage_map_key::<AccountId>("System", "Account", address.clone())?;
-
-		info!("storage key is: 0x{}", hex::encode(&storagekey));
-		self.get_storage_by_key_hash(storagekey, None)
-	}
-
-	pub fn get_account_data(
-		&self,
-		address: &Runtime::AccountId,
-	) -> ApiResult<Option<Runtime::AccountData>> {
-		self.get_account_info(address).map(|info| info.map(|i| i.data))
-	}
-
-	pub fn get_finalized_head(&self) -> ApiResult<Option<Runtime::Hash>> {
-		let h = self.get_request(json_req::chain_get_finalized_head())?;
-		match h {
-			Some(hash) => Ok(Some(Runtime::Hash::from_hex(hash)?)),
-			None => Ok(None),
-		}
-	}
-
-	pub fn get_header<H>(&self, hash: Option<Runtime::Hash>) -> ApiResult<Option<H>>
-	where
-		H: Header + DeserializeOwned,
-	{
-		let h = self.get_request(json_req::chain_get_header(hash))?;
-		match h {
-			Some(hash) => Ok(Some(serde_json::from_str(&hash)?)),
-			None => Ok(None),
-		}
-	}
-
-	pub fn get_block_hash(
-		&self,
-		number: Option<Runtime::BlockNumber>,
-	) -> ApiResult<Option<Runtime::Hash>> {
-		let h = self.get_request(json_req::chain_get_block_hash(number))?;
-		match h {
-			Some(hash) => Ok(Some(Runtime::Hash::from_hex(hash)?)),
-			None => Ok(None),
-		}
-	}
-
-	pub fn get_block<B>(&self, hash: Option<Runtime::Hash>) -> ApiResult<Option<B>>
-	where
-		B: Block + DeserializeOwned,
-	{
-		Self::get_signed_block(self, hash).map(|sb_opt| sb_opt.map(|sb| sb.block))
-	}
-
-	pub fn get_block_by_num<B>(&self, number: Option<Runtime::BlockNumber>) -> ApiResult<Option<B>>
-	where
-		B: Block + DeserializeOwned,
-	{
-		Self::get_signed_block_by_num(self, number).map(|sb_opt| sb_opt.map(|sb| sb.block))
-	}
-
-	/// A signed block is a block with Justification ,i.e., a Grandpa finality proof.
-	/// The interval at which finality proofs are provided is set via the
-	/// the `GrandpaConfig.justification_period` in a node's service.rs.
-	/// The Justification may be none.
-	pub fn get_signed_block<B>(
-		&self,
-		hash: Option<Runtime::Hash>,
-	) -> ApiResult<Option<SignedBlock<B>>>
-	where
-		B: Block + DeserializeOwned,
-	{
-		let b = self.get_request(json_req::chain_get_block(hash))?;
-		match b {
-			Some(block) => Ok(Some(serde_json::from_str(&block)?)),
-			None => Ok(None),
-		}
-	}
-
-	pub fn get_signed_block_by_num<B>(
-		&self,
-		number: Option<Runtime::BlockNumber>,
-	) -> ApiResult<Option<SignedBlock<B>>>
-	where
-		B: Block + DeserializeOwned,
-	{
-		self.get_block_hash(number).map(|h| self.get_signed_block(h))?
-	}
-
 	pub fn get_request(&self, jsonreq: Value) -> ApiResult<Option<String>> {
 		self.client.get_request(jsonreq).map_err(ApiClientError::RpcClient)
-	}
-
-	pub fn get_storage_value<V: Decode>(
-		&self,
-		storage_prefix: &'static str,
-		storage_key_name: &'static str,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<V>> {
-		let storagekey = self.metadata.storage_value_key(storage_prefix, storage_key_name)?;
-		info!("storage key is: 0x{}", hex::encode(&storagekey));
-		self.get_storage_by_key_hash(storagekey, at_block)
-	}
-
-	pub fn get_storage_map<K: Encode, V: Decode>(
-		&self,
-		storage_prefix: &'static str,
-		storage_key_name: &'static str,
-		map_key: K,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<V>> {
-		let storagekey =
-			self.metadata.storage_map_key::<K>(storage_prefix, storage_key_name, map_key)?;
-		info!("storage key is: 0x{}", hex::encode(&storagekey));
-		self.get_storage_by_key_hash(storagekey, at_block)
-	}
-
-	pub fn get_storage_map_key_prefix(
-		&self,
-		storage_prefix: &'static str,
-		storage_key_name: &'static str,
-	) -> ApiResult<StorageKey> {
-		self.metadata
-			.storage_map_key_prefix(storage_prefix, storage_key_name)
-			.map_err(|e| e.into())
-	}
-
-	pub fn get_storage_double_map<K: Encode, Q: Encode, V: Decode>(
-		&self,
-		storage_prefix: &'static str,
-		storage_key_name: &'static str,
-		first: K,
-		second: Q,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<V>> {
-		let storagekey = self.metadata.storage_double_map_key::<K, Q>(
-			storage_prefix,
-			storage_key_name,
-			first,
-			second,
-		)?;
-		info!("storage key is: 0x{}", hex::encode(&storagekey));
-		self.get_storage_by_key_hash(storagekey, at_block)
-	}
-
-	pub fn get_storage_by_key_hash<V: Decode>(
-		&self,
-		key: StorageKey,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<V>> {
-		let s = self.get_opaque_storage_by_key_hash(key, at_block)?;
-		match s {
-			Some(storage) => Ok(Some(Decode::decode(&mut storage.as_slice())?)),
-			None => Ok(None),
-		}
-	}
-
-	pub fn get_opaque_storage_by_key_hash(
-		&self,
-		key: StorageKey,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<Vec<u8>>> {
-		let jsonreq = json_req::state_get_storage(key, at_block);
-		let s = self.get_request(jsonreq)?;
-
-		match s {
-			Some(storage) => Ok(Some(Vec::from_hex(storage)?)),
-			None => Ok(None),
-		}
-	}
-
-	pub fn get_storage_value_proof(
-		&self,
-		storage_prefix: &'static str,
-		storage_key_name: &'static str,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<ReadProof<Runtime::Hash>>> {
-		let storagekey = self.metadata.storage_value_key(storage_prefix, storage_key_name)?;
-		info!("storage key is: 0x{}", hex::encode(&storagekey));
-		self.get_storage_proof_by_keys(vec![storagekey], at_block)
-	}
-
-	pub fn get_storage_map_proof<K: Encode, V: Decode + Clone>(
-		&self,
-		storage_prefix: &'static str,
-		storage_key_name: &'static str,
-		map_key: K,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<ReadProof<Runtime::Hash>>> {
-		let storagekey =
-			self.metadata.storage_map_key::<K>(storage_prefix, storage_key_name, map_key)?;
-		info!("storage key is: 0x{}", hex::encode(&storagekey));
-		self.get_storage_proof_by_keys(vec![storagekey], at_block)
-	}
-
-	pub fn get_storage_double_map_proof<K: Encode, Q: Encode, V: Decode + Clone>(
-		&self,
-		storage_prefix: &'static str,
-		storage_key_name: &'static str,
-		first: K,
-		second: Q,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<ReadProof<Runtime::Hash>>> {
-		let storagekey = self.metadata.storage_double_map_key::<K, Q>(
-			storage_prefix,
-			storage_key_name,
-			first,
-			second,
-		)?;
-		info!("storage key is: 0x{}", hex::encode(&storagekey));
-		self.get_storage_proof_by_keys(vec![storagekey], at_block)
-	}
-
-	pub fn get_storage_proof_by_keys(
-		&self,
-		keys: Vec<StorageKey>,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<ReadProof<Runtime::Hash>>> {
-		let jsonreq = json_req::state_get_read_proof(keys, at_block);
-		let p = self.get_request(jsonreq)?;
-		match p {
-			Some(proof) => Ok(Some(serde_json::from_str(&proof)?)),
-			None => Ok(None),
-		}
-	}
-
-	pub fn get_keys(
-		&self,
-		key: StorageKey,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<Vec<String>>> {
-		let jsonreq = json_req::state_get_keys(key, at_block);
-		let k = self.get_request(jsonreq)?;
-		match k {
-			Some(keys) => Ok(Some(serde_json::from_str(&keys)?)),
-			None => Ok(None),
-		}
-	}
-
-	pub fn get_fee_details(
-		&self,
-		xthex_prefixed: &str,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<FeeDetails<Runtime::Balance>>> {
-		let jsonreq = json_req::payment_query_fee_details(xthex_prefixed, at_block);
-		let res = self.get_request(jsonreq)?;
-		match res {
-			Some(details) => {
-				let details: FeeDetails<NumberOrHex> = serde_json::from_str(&details)?;
-				let details = convert_fee_details(details)?;
-				Ok(Some(details))
-			},
-			None => Ok(None),
-		}
-	}
-
-	pub fn get_payment_info(
-		&self,
-		xthex_prefixed: &str,
-		at_block: Option<Runtime::Hash>,
-	) -> ApiResult<Option<RuntimeDispatchInfo<Runtime::Balance>>> {
-		let jsonreq = json_req::payment_query_info(xthex_prefixed, at_block);
-		let res = self.get_request(jsonreq)?;
-		match res {
-			Some(info) => {
-				let info: RuntimeDispatchInfo<Runtime::Balance> = serde_json::from_str(&info)?;
-				Ok(Some(info))
-			},
-			None => Ok(None),
-		}
 	}
 
 	pub fn get_constant<C: Decode>(
@@ -595,10 +325,6 @@ where
 			.ok_or(MetadataError::ConstantNotFound(constant))?;
 
 		Ok(Decode::decode(&mut c.value.as_slice())?)
-	}
-
-	pub fn get_existential_deposit(&self) -> ApiResult<Runtime::Balance> {
-		self.get_constant("Balances", "ExistentialDeposit")
 	}
 
 	#[cfg(feature = "ws-client")]
@@ -619,29 +345,4 @@ where
 		// XtStatus should never be used used but we need to put something
 		self.client.send_extrinsic(xthex_prefixed, XtStatus::Broadcast)
 	}
-}
-
-fn convert_fee_details<Balance: TryFrom<NumberOrHex>>(
-	details: FeeDetails<NumberOrHex>,
-) -> ApiResult<FeeDetails<Balance>> {
-	let inclusion_fee = if let Some(inclusion_fee) = details.inclusion_fee {
-		Some(inclusion_fee_with_balance(inclusion_fee)?)
-	} else {
-		None
-	};
-	let tip = details.tip.try_into().map_err(|_| ApiClientError::TryFromIntError)?;
-	Ok(FeeDetails { inclusion_fee, tip })
-}
-
-fn inclusion_fee_with_balance<Balance: TryFrom<NumberOrHex>>(
-	inclusion_fee: InclusionFee<NumberOrHex>,
-) -> ApiResult<InclusionFee<Balance>> {
-	Ok(InclusionFee {
-		base_fee: inclusion_fee.base_fee.try_into().map_err(|_| ApiClientError::TryFromIntError)?,
-		len_fee: inclusion_fee.len_fee.try_into().map_err(|_| ApiClientError::TryFromIntError)?,
-		adjusted_weight_fee: inclusion_fee
-			.adjusted_weight_fee
-			.try_into()
-			.map_err(|_| ApiClientError::TryFromIntError)?,
-	})
 }
